+++ src/components/ClickOutHandler.tsx
@@ -1,22 +1,42 @@
-import React, { useEffect, useRef } from 'react'
+import React, { ReactNode, useCallback, useEffect, useRef } from 'react'
 
 type Props = {
   // eslint-disable-next-line no-empty-pattern
-  children: React.ReactNode | (({}) => any)
+  children: ReactNode | (({}) => any)
   enabled?: boolean
   events?: string[]
   ignoredElements?: HTMLElement[]
-  refProp?: string
-  wrapWith?: React.ElementType
   onClickOut: (ev: Event) => any
 }
 
-const ClickOutHandler = (props: Props) => {
+const ClickOutHandler = ({
+  children,
+  enabled = true,
+  events = ['mousedown', 'touchstart'],
+  ignoredElements = [],
+  onClickOut,
+}: Props) => {
   const wrapperRef = useRef<HTMLElement | null>(null)
 
-  useEffect(() => {
-    const events = props.events || ['mousedown', 'touchstart']
+  const shouldFire = useCallback(
+    (ev: Event) => {
+      return (
+        enabled &&
+        wrapperRef.current &&
+        !wrapperRef.current.contains(ev.target as HTMLElement) &&
+        !ignoredElements.some((element) => element && element.contains(ev.target as HTMLElement))
+      )
+    },
+    [enabled, ignoredElements],
+  )
 
+  useEffect(() => {
+    console.log('fired')
+    const handleClickOut = (ev: Event) => {
+      if (shouldFire(ev)) {
+        onClickOut(ev)
+      }
+    }
     events.forEach((event) => {
       document.addEventListener(event, handleClickOut)
     })
@@ -26,40 +46,15 @@ const ClickOutHandler = (props: Props) => {
         document.removeEventListener(event, handleClickOut)
       })
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  const handleClickOut = (ev: Event) => {
-    if (shouldFire(ev)) {
-      props.onClickOut(ev)
-    }
-  }
-
-  const shouldFire = (ev: Event) => {
-    return (
-      props.enabled !== false &&
-      wrapperRef.current &&
-      !wrapperRef.current.contains(ev.target as any) &&
-      !(props.ignoredElements || []).some((element) => element && element.contains(ev.target as any))
-    )
-  }
-
-  const { children, refProp = 'ref', wrapWith } = props
-  const passRef = (el: HTMLElement) => {
-    wrapperRef.current = el
-  }
+  }, [events, onClickOut, shouldFire])
 
   if (typeof children === 'function') {
-    return children({ [refProp]: passRef })
+    return children({ ref: wrapperRef })
   }
 
-  if (Array.isArray(children) || wrapWith) {
-    const Wrapped = wrapWith || 'div'
-
-    return <Wrapped {...{ [refProp]: passRef }}>{children}</Wrapped>
-  }
-
-  return React.cloneElement(React.Children.only(children as any), { [refProp]: passRef })
+  return React.cloneElement(React.Children.only(children as any), {
+    ref: wrapperRef,
+  })
 }
 
 export default ClickOutHandler